write GB palettes through subroutine
The first step of GBC enhancement is to make the program do the same
thing on a GB and a GBC.  The first step toward this is to refactor
all writes to rBGP or rOBPx to instead call a subroutine that
performs that write and then performs equivalent writes to the GBC
palette registers.  Actual effect on GBC will be tested later;
this commit is for setting up the infrastructure.