compress mono tilemaps as incruniq (.iu)
It's not quite a miracle, but turning runs of map entries
representing singleton tiles into runs before pb16 on the tilemap
saved about 100 bytes.  I should be able to eke out more by
compressing the GBC version of the Gus portrait, after which I'll
have 1K which is enough to start writing the menu for the GBC tests.