GBA sound test: Play synthesized PCM chord
Use the Karplus/Strong algorithm (short wideband excitation echoed
with a period in the audio frequency) to synthesize six plucked
strings playing a G major chord, mix them, and play them on one
of the two PCM channels.

Code in IWRAM runs faster than code in EWRAM or ROM, even if it's
Thumb code.

We're very close to a release of 0.02.