add reference copy of fireshark image
The Pluge Contrast test, new in 240p Test Suite (for Genesis) 1.16,
includes a 32x32 pixel, 7 color "FireShark" icon used as a tiled
background.  It's suitable for GBA but would need compromises on
GB and NES.  Normally, to guide the compromise, I'd inquire about
its authorship, but lately, I may be coming off as demanding on the
Shmups Forum topic about 240p Test Suite.  So I'll lie low for a
while and exclude the Pluge Contrast test from the immediate next
version of the NES, GB, and GBA suite.

Should the author prove to be a major proprietary game publisher
or otherwise hostile, revert this commit.  In that case, I'll
have to come up with some other emblem with comparable metrics.