GBA: align help line buffer
DMA memset works only with aligned buffers.  A DMA that isn't aligned
may corrupt nearby bytes.  In this case, the vertical scroll test's
autotiler clears and uses the help page's scratch buffer, and failure
to align the buffer caused DMA to corrupt other variables, in this
case the height in lines of the loaded help page.  This in turn
caused the Down arrow not to work because the cursor was already
below what it thought was the bottom of the page.