compress 2-bit shark
This is the start of implementing something similar to PLUGE Contrast
(issue #6) on NES and GBC.  I expect the GBC version to look similar
to the GBA version, while the NES version will probably be coarser
than the Genesis original.  The compressed tile data is included in
the ROM, but not yet any code to display it.