NES: use iu53 where appropriate
Converting Sharpness, Stopwatch, and Crosstalk to iu53 compression
regains the space lost to the more detailed picture of Gus on the
help screen.  Adding a command line argument to sav2iu53 to specify a
tile other than solid color 0 as the blank tile lets me convert the
portrait of Gus in Shadow sprite.  Now we have 22.5 KiB free and only
288 bytes used in bank 1, all in the "gatedata" subsystem.

This emboldens me to shoot for 24 KiB free, but first I'd need to run
activity-by-activity comparative space analysis vs. the GBC version.