GB VWF: allocate HRAM with RSSET/RB not EQU/SET
ISSOtm in gbdev Discord claims that using anything but UNION to
allocate local variables in HRAM is bad practice.  But using UNION in
an RGBDS project with more than one translation unit has a serious
drawback: a subroutine and its locals have to be defined in separate
files.  This is true whether all subroutines' locals live in one
file, or each subroutine's locals are in a separate file that the
UNION file INCLUDEs.

I acknowledge that defining locals with SET is a bad practice.
Allocating them with RSSET and RB, as this commit does, is less bad.
But it's still not perfect when both a caller and a callee have local
variables, as the callee may later be changed to use more locals than
before, clobbering the caller's locals.  A complete solution for both
ca65 and RGBDS would move local allocation to a preprocessor.