reduce magic numbers in audiosync
ISSOtm made pull request #11 that saved a byte by assuming that
the bottom of the ceiling is on a 4-row boundary.  But this
might change if Artemio gets around to unifying the ceiling's
size between the Super NES and Genesis versions of this test.
So I turned a bunch of distances and durations into constants
changeable at the top of the file.

I did find a redundant load, however.