linearity: compress with reflection
The bottom half of Linearity is the top half reflected.
The right half of Linearity is the left half reflected.
Doing the reflection at runtime means we can store less tile data
and less map data.  (It's done in software to keep it working
on DMG/SGB and keep the grid working.)

This change expands unused space from 329 to 898 bytes,
saving 569 bytes.