GBA: apply .grit to graphics in common/tilesets
Each PNG image in gba/tilesets has a corresponding file with .grit
extension listing the options for converting that image.  Normally
grit looks for an options file in the same folder as the image, but
this can be overridden.  It wasn't found for pluge_shark_6color.png
because the PNG file was in common/tilesets, not gba/tilesets,
which caused grit to convert that image using default settings.
This broke everything when the player selects the PLUGE test.

So I had the makefile pass the options file's name to grit using
-ff/path/to/optionsfile.grit