First border proposal
SuperFamiconv reports 63 unique tiles in this border, compressing
to 941 PB16 bytes.  The tilemap has 8 unique rows of 22 tiles.
With only 64 tiles, we can move VH and palette to bits 7-6, for a
total of 8*22 + 28 = 204 bytes of tilemap.  That's 1145 bytes plus
however many to unpack them and push them to the SGB.  But as of
today, only 336 bytes are free in the ROM.  Thus actually adding
this border to the ROM is pending another compression breakthrough.