add experimental VWF packing estimator
With the larger RAM of the Game Boy, we can pack the VWF tiles
in ROM and decompress it to WRAM at startup time.  Not storing
leading or trailing blank rows or trailing blank columns
compresses the 104 tiles of common/tilesets/vwf7_cp144p.png
from 936 to 459 bytes saving 477.  However, the unpacking code
for this format would eat a lot of the savings.

Also add a tool to visualize the entire ROM in order to find
anything distinguishable from random noise.