GB: Nibble Huffman coding experiment
I started to write a nibble-oriented Huffman compressor.  It packs
the VWF data from 832 bytes to 507 in addition to saving at least
1100 bytes over a subset of PB16 data.  This alone might be enough
to fit the border, so long as the unpacker isn't inordinately slow.
The big win over the other rectangular packer proposal is sharing
the unpacker with PB16 data.