GB Shadow sprite: Gus to tilemap 2; no more 8x16
For #21, I want to have all tilemaps in VRAM at once.  This
requires rearranging the tilemap for horizontal and diagonal
striped patterns to draw the last four rows with the window.
That means the frame type display may require using sprites
instead of the window, and if so, it'll have to be 8x8 because
that's all the VWF library supports.