Huffman may be too bottlenecky
Begin to analyze practicality of a Huffman tree with a maximum
node-to-node offset of 4 nodes.  In a fairly balanced tree,
I expect congestion at depth three: things quickly become
congested with no room for more internal nodes.  Example of a
tree with 14 internal nodes (L represents a leaf):

                0
            /       \
        1               4
      /   \           /   \
    2       5       6       8
   / \     / \     / \     / \
  L   3   7   9   L  10  11  12

The top of this is a perfectly balanced binary tree, which
may allows a separate set of index spaces:

  |   |   |   |   |   |   |   |
  L  A0  B0  C0   L  D0  E0  F0

But even with this, allocating internal nodes is tricky
especially for a tree that's unbalanced at the top (like unary)
but has a balanced part deeper down, and I doubt it's simpler
than Canonical Huffman.

For #21