GB: refactor gbcnamtool; add pb16.py -d
To make the Huffman file directory thing for #21, I'll need to be
able to decompress PB16 or at least find the ending offset after
a number of packets.  So add decompression support to PB16.

I may also need to change the file format of gbcnamtool's output
from assembly to binary in order to include GBC IU files.