GB: jr *+3 to db $FE
Instead of unconditionally branching over a 1-byte opcode with
JR, turn the branch into a CP with the other opcode as its
operand to save 1 byte in each of 3 cases.  Add a test for this
to tools/unused.py, avoiding two classes of false alarms.

suggested by @ISSOtm in gbdev Discord