GB: DTE reencoding
After I added number of occurrences to the verbose output, I found
an anomaly in the compressed data for the "Linearity" section:
"Game Boy " wasn't represented with the shortest possible sequence.
So I wrote a reencoder using a greedy algorithm, which saved 28
bytes over @jroatch's encoder.  I now use his encoder to make the
replacements table and then actually encode the text in Python.

Unused is up to 675.