GB DTE: choose greedy line-by-line
After I explained the improvements from the greedy reencoder to
@jroatch, he showed me this counterexample:

    Say DTE produces the dictionary [bc, de, abc, def, bcde]
    and you're encoding the string "abcdef".
    Greedy would find "bcde" and produce 3 bytes a|bcde|f
    while an optimal algorithm would produce 2 bytes abc|def.

After a help text change to reflect row inversion behavior, I found
a similar counterexample in the Overscan text.  To work around this,
pending development of an optimal encoder, break up the text at
newlines and choose which is smaller for each line.