GBA: -fno-common for forthcoming GCC 10
GCC 9 and earlier default to -fcommon, which puts global variables
defined in multiple translation units into a "common" section where
the linker sorts out duplicates.  The -fcommon feature is intended
for Fortran and for pre-standard C, which had this behavior.
GCC 10 and later default to -fno-common, which reflects standard C
behavior in which this causes a linker error.

Reported by ax6 against a different project
https://github.com/rednex/rgbds/pull/504#issuecomment-610438094