GB Help: don't move cursor to blank line
Previously, if the last line of a page was a repeated helptitle,
the help system was seeing the page as ending with a single blank
line before the NUL terminator That'd be fine, except if it was a
menu, it would let the user move the cursor onto the blank line.
Look one character ahead for the NUL terminator to prevent this.