add NES manual and GPLv2 pamphlet
In case someone wants to make a physical release, I've started
on a manual that prints on both sides of a sheet of US Letter
size paper.  I made a 1-sheet layout of the GNU General Public
License version 2 as well.  Writing the manual involved
copyediting the intro and controls sections of README.md a bit
for consistency.

Fonts used: Liberation Sans and Liberation Sans Narrow