GB: fix duplicate gfxincbins section names
New RGBDS introduces explicit behaviors to apply when SECTIONs
share a name.

- SECTION UNION overlaps each section in memory.  Useful for local
  variables or for state that pertains to only one game phase.
- SECTION FRAGMENT concatenates the contents, ensuring they end up
  in the same BANK.

To encourage users to document their intent, it adds a warning for
SECTIONs that share a name but specify no behavior, suggesting use
of SECTION FRAGMENT.  A copy-and-paste typo in 144p Test Suite
resulted in two SECTIONs in gfxincbins.z80 that shared a name.
Fix the typo.