GB: golf decompression to CHRRAM0
The most common destination for PB16 decompression is CHRRAM0
($8000).  I made a special case fall-through for this address.
At the same time, I consolidated CHR loading routines called by
multiple activities, such as common tiles (bggfx) and SMPTE tiles,
to turn JP into JR.

read_pad_help_check was a bust, so I'll close #29 with a gain
of 67 bytes.  This is roughly the entire size of the setup-less
SMPTE bars (#28).