GB: begin measuring BSS
At the beginning of a program in the C language, uninitialized
variables have the value 0.  C implementations commonly put
these variables in a section called BSS and memset() it to 0
before calling the main function.  RGBDS, however, includes no
way to measure the start and size of a section including all its
fragments.  So in the gbdev Discord server, ISSOtm suggested
adding files linked first and last to establish the start and end.
I expect to use this soon when size-optimizing variable zeroing.

At the same time, I upgraded RGBDS, which exposed that it is now
more sensitive to lack of a newline at the end of an included file.