begin work on rear view
Persune on 240p Test Suite Discord server suggested drawing Gus
as a roly-poly toy in a forthcoming Donna poster.  In-universe,
that could mean the TV shop Gus works for might give out toys
as premiums to customers.  This means I need to texture and
sculpt the front and back of an injection-molded toy.