Actions: don't apt-get upgrade
Phased updates of anything GRUB-adjacent in Ubuntu 22.04 LTS and
later run at a snail's pace, taking a month to finish.  This is
causing apt-get -y upgrade in GitHub Actions recipes to fail.
https://askubuntu.com/q/1431786/232993

I have no idea why we're trying to update GRUB on an Actions Azure
VM anyway.  I discussed this with Persune and we ended up deciding
to drop the upgrade step, doing only the update+install.